Updated Player SDK version to 1.6.0;
