+++ app/build.gradle
@@ -45,17 +45,18 @@ android {
 
     dependencies {
         // Android
-        implementation 'androidx.core:core-ktx:1.7.0-beta02'
-        implementation 'androidx.appcompat:appcompat:1.3.1'
-        implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
+        api 'androidx.core:core-ktx:1.7.0'
+        implementation 'androidx.appcompat:appcompat:1.4.0'
+        implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
         implementation 'com.google.android.material:material:1.4.0'
 
         // Kotlin
-        implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.2.0'
+        implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.1'
 
-        // Lifecycle
-        implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.1'
-        implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1'
+        // Coroutines
+        implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.4.0'
+        implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2"
+        implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2"
 
         // Timber
         implementation 'com.jakewharton.timber:timber:5.0.1'
@@ -67,6 +68,6 @@ android {
         implementation 'androidx.exifinterface:exifinterface:1.3.3'
 
         // Amazon IVS
-        implementation 'com.amazonaws:ivs-player:1.5.1'
+        implementation 'com.amazonaws:ivs-player:1.6.0'
     }
 }
