+++ app/src/main/java/com/amazonaws/ivs/player/scrollablefeed/common/CoroutineExtensions.kt
@@ -1,0 +1,35 @@
+package com.amazonaws.ivs.player.scrollablefeed.common
+
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import kotlinx.coroutines.*
+import kotlinx.coroutines.channels.BufferOverflow
+import kotlinx.coroutines.flow.MutableSharedFlow
+import timber.log.Timber
+
+private val mainScope = CoroutineScope(Dispatchers.Main + SupervisorJob())
+
+@Suppress("FunctionName")
+fun <T> ConsumableSharedFlow(canReplay: Boolean = false) = MutableSharedFlow<T>(
+    replay = if (canReplay) 1 else 0,
+    extraBufferCapacity = 1,
+    onBufferOverflow = BufferOverflow.DROP_OLDEST
+)
+
+fun launchMain(block: suspend CoroutineScope.() -> Unit) = mainScope.launch(
+    context = CoroutineExceptionHandler { _, e -> Timber.w(e, "Coroutine failed: ${e.localizedMessage}") },
+    block = block
+)
+
+fun AppCompatActivity.launchUI(
+    lifecycleState: Lifecycle.State = Lifecycle.State.STARTED,
+    block: suspend CoroutineScope.() -> Unit
+) = lifecycleScope.launch(
+    context = CoroutineExceptionHandler { _, e ->
+        Timber.d(e, "Coroutine failed: ${e.localizedMessage}")
+    }
+) {
+    repeatOnLifecycle(state = lifecycleState, block = block)
+}
