+++ app/src/main/java/com/amazonaws/ivs/player/scrollablefeed/common/ViewExtensions.kt
@@ -2,6 +2,7 @@ package com.amazonaws.ivs.player.scrollablefeed.common
 
 import android.content.Context
 import android.graphics.Bitmap
+import android.graphics.SurfaceTexture
 import android.os.Build
 import android.os.Handler
 import android.os.Looper
@@ -17,7 +18,14 @@ import java.util.*
 import java.util.concurrent.TimeUnit
 import kotlin.math.abs
 import android.graphics.drawable.BitmapDrawable
+import android.text.Spannable
+import android.text.SpannableStringBuilder
+import android.text.style.ForegroundColorSpan
+import android.util.Size
+import androidx.core.view.doOnLayout
+import com.amazonaws.ivs.player.scrollablefeed.models.SizeModel
 import kotlinx.coroutines.delay
+import kotlin.math.roundToInt
 
 fun initBlurredBackground(textureView: TextureView, background: ImageView) = launchMain {
     if (background.hasImage()) return@launchMain
@@ -85,6 +93,12 @@ fun Bitmap.scaleBitmap(maxWidth: Int, maxHeight: Int): Bitmap {
     return Bitmap.createScaledBitmap(this, width, height, true)
 }
 
+fun TextView.firstCharColor(color: Int) {
+    val spannable = SpannableStringBuilder(text)
+    spannable.setSpan(ForegroundColorSpan(color), 0, 1, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
+    text = spannable
+}
+
 fun TextView.setActiveTime(time: String, currentTime: Long) {
     val message = StringBuilder("for")
     val pattern = "yyyy-MM-dd hh:mm:ss"
@@ -104,6 +118,55 @@ fun TextView.setActiveTime(time: String, currentTime: Long) {
         message.append(" ${minutes}m")
         text = message
     }
+}
+
+fun TextureView.onReady(onReady: (surface: Surface) -> Unit) {
+    if (surfaceTexture != null) {
+        onReady(Surface(surfaceTexture))
+        return
+    }
+    surfaceTextureListener = object : TextureView.SurfaceTextureListener {
+        override fun onSurfaceTextureAvailable(surface: SurfaceTexture, width: Int, height: Int) {
+            surfaceTextureListener = null
+            onReady(Surface(surfaceTexture))
+        }
+
+        override fun onSurfaceTextureSizeChanged(surface: SurfaceTexture, width: Int, height: Int) {
+            /* Ignored */
+        }
+
+        override fun onSurfaceTextureDestroyed(surface: SurfaceTexture) = false
+
+        override fun onSurfaceTextureUpdated(surface: SurfaceTexture) {
+            /* Ignored */
+        }
+    }
+}
+
+fun View.scaleToFit(videoSize: SizeModel, parentView: View? = null) {
+    if (videoSize.width == 0 || videoSize.height == 0) return
+    (parentView ?: parent as View).doOnLayout { useToScale ->
+        calculateSurfaceSize(videoSize.width, videoSize.height)
+        val size = useToScale.calculateSurfaceSize(videoSize.width, videoSize.height)
+        val params = layoutParams
+        params.width = size.width
+        params.height = size.height
+        layoutParams = params
+    }
+}
+
+private fun View.calculateSurfaceSize(videoWidth: Int, videoHeight: Int): Size {
+    val ratio = videoHeight / videoWidth.toFloat()
+    val scaledWidth: Int
+    val scaledHeight: Int
+    if (measuredHeight > measuredWidth * ratio) {
+        scaledWidth = measuredWidth
+        scaledHeight = (measuredWidth * ratio).roundToInt()
+    } else {
+        scaledWidth = (measuredHeight / ratio).roundToInt()
+        scaledHeight = measuredHeight
+    }
+    return Size(scaledWidth, scaledHeight)
 }
 
 private fun ImageView.hasImage(): Boolean {
