+++ app/src/main/java/com/amazonaws/ivs/player/scrollablefeed/models/StreamUIModel.kt
@@ -1,6 +1,6 @@
 package com.amazonaws.ivs.player.scrollablefeed.models
 
-data class StreamViewModel(
+data class StreamUIModel(
     val id: Int,
     val stream: StreamItemModel,
     val metadata: MetadataModel,
