+++ app/src/main/res/layout/stream_item.xml
@@ -7,7 +7,7 @@
         <import type="android.view.View" />
         <variable
             name="stream"
-            type="com.amazonaws.ivs.player.scrollablefeed.models.StreamViewModel" />
+            type="com.amazonaws.ivs.player.scrollablefeed.models.StreamUIModel" />
         <variable
             name="isCenter"
             type="Boolean" />
