+++ app/src/main/res/layout/stream_title_view.xml
@@ -7,7 +7,7 @@
         <import type="android.view.View" />
         <variable
             name="stream"
-            type="com.amazonaws.ivs.player.scrollablefeed.models.StreamViewModel" />
+            type="com.amazonaws.ivs.player.scrollablefeed.models.StreamUIModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
